Add Travis CI File
Added a .travis.yml file to (hopefully) enable build testing for various arduino platforms on this repository. Travis may not work with this commit, as I have virtually no experience with Travis, and it may take further adjustment to achieve the desired result. Please also note that this will ONLY test if the library successfully builds, and does NOT have unit tests, or any other mechanism to access the functionality. Use caution.